extract pipeline tests using translator
