+++ e2e/document_stores/test_cosine_similarity.py
@@ -6,7 +6,7 @@ import numpy as np
 
 from haystack.schema import Document
 
-from .conftest import document_store
+from ..conftest import document_store
 
 
 DOCUMENTS = [
