+++ e2e/nodes/test_translator.py
